Inline IMMS#startInputOrWindowGainedFocusInternal
This is a follow up CL to my previous CL [1], which aimed to be a
logical revert of changes mentioned in the commit message.

One thing that CL forgot to handle was

  InputMethodManagerService#startInputOrWindowGainedFocusInternal,

which was mentioned in one of our previous CLs [2].

This CL inlines that method again for simplicity.

This is a mechanical refactoring.  There should be no behavior change.

 [1]: If16ac0de536d9089eb04f6e07b1ee47378124658
      c617b994210efa94408a085ec1c5c5821ca147e3
 [2]: Ic584203c1221fbae17f5e2d8f09e3992df061646
      4d73c3e390ba6218e9188a7d13666d238d2c472d

Bug: 234882948
Test: presubmit
Change-Id: I7c775245882815483f29acfd83f53b08a40d07ed