Clear default focus highlight when lost window focus or re-enter touch mode
The default focus highlight will indicate the focused view in view
system when user navigate the focus without touch.

This CL will clear it if the window focus has lost or it re-entered the
touch mode, so we won't see the wrong highlights in different windows
while using multi-window mode.

Bug: 244273901
Test: atest DefaultFocusHighlightTest
Test: multi-window, use dpad to navigate focus
Change-Id: Iad07c77b4ae16f3c88797eddd749e0aedebdd4fc