Fix alpha not fading on notifications after the first shade expansion
The problem has regressed due to b/239903210 fix. During the development, we have made a refactoring where we centralized all the `alpha` handling logic. While centralising the code, we inlined a function which had an early return statement and caused the code execution to stop before expected. By changing the early return to a `continue` we ensure the fade works correctly.

Test: manual

Test: atest ~/tm-qpr-dev/frameworks/base/packages/SystemUI/tests/src/com/android/systemui/statusbar/notification/stack/StackScrollAlgorithmTest.kt

Fixes: 245506240

Change-Id: Iee6c2ed0e5ef83e5cb659ee0363cbd80a3c2166e