Use uptime for activity startup time
It was changed from uptime to elapsed realtime in [1] because when
changing the time unit from millisecond to nanosecond, the method
uptimeNanos is not yet available.

It may cause unexpected long duration if someone starts activity
without holding wakelock while device is not awake and the device
enters deep sleep state in the middle of the startup.

Since uptimeNanos is added, switch to it to avoid the noise.

[1]: I75c0be8508499dce186fd392e79b67b2fefc8842

Bug: 204488635
Test: atest ActivityMetricsLoggerTests

Change-Id: I86db562b6dbffa4565d21c2942b258af8e3bc44e