Use simple home key for setup
The step in setup is not the major action to verify. And because
tapl may use its context configuration rotation to assert (may
not be updated in time), that causes some unnecessary failures.

The test is to verify the animation when launching from notification,
so how to return to home in setup is not so important.

Bug: 244414110
Test: atest FlickerTestsOpen:AppFromLockNotificationWarm
Change-Id: I94382fdab3beb92d3740bd1850ee2ae7f4d8d0c9